adding edit fragment
